Move test_s3_roundtrip from test_s3.py to test_parquet.py
